requirements.txt
django==4.0.5
django-crispy-forms
crispy-bootstrap5
django-ckeditor
PyPDF2<3.0
Pillow
requests